Refactor CLI from 2046-line monolith into clean package
Split src/cli.py into src/cli/ package with 7 focused modules:
- config.py: path constants, providers, Docker helpers, agent management
- formatting.py: tool display, styled output, response rendering
- channels.py: ChannelManager for messaging channel lifecycle
- runtime.py: RuntimeContext for full runtime lifecycle management
- repl.py: REPLSession with command registry pattern
- main.py: thin Click command definitions
- __init__.py: backward-compatible re-exports

Key improvements:
- Eliminate 3x duplicated tool-output display code
- Styled startup output with section headers and status indicators
- REPL prompt shows active agent: [assistant] You>
- Error tool hints displayed in red instead of dim gray
- Channel status properly ordered under Channels header
- RuntimeContext class replaces 372-line _start_interactive()
- REPLSession replaces 257-line if/elif command dispatch chain
- dispatch() sync method fixed (was returning coroutine)

Updates test patches to target src.cli.config.* and CLAUDE.md
to reflect new module structure. All 717 tests pass.
