+++ src/cli/__main__.py
@@ -1,0 +1,5 @@
+"""Allow `python -m src.cli` to work."""
+
+from src.cli.main import cli
+
+cli()
